Refactor into single file
